Break settings into more components
